Add files via upload
a copy of the all human ratings — there are 60 articles, 7 systems, and 3 ratings by different people for each system for a total of 1,260 rows. Each row is a single rating across the four dimensions from the paper. The columns are:

System - name of the sumarization system/baseline being rated
ArticleID - index of the article in the dataset, can be used to group by article
ArticleText, ArticleTitle, SystemSummary - three inputs the rater sees (HTML encoded for AMT)
CoherenceRating, FluencyRating, InformativenessRating, RelevanceRating - ratings (1-5)